Clarify instructions and add recommendations
The first section which introduces the core concepts behind this
document has the most changes, all of which were for clarification
and readability. Warnings and recommendations were added throughout
the instructions to recommend using the CS lab machines, check that
their private repository is empty on creation, and to frequently
commit local changes and frequently fetch changes on the upstream.